Create common folder and move examples
