+++ frontend/src/common/examples/BasicExample.tsx
