+++ frontend/src/common/examples/ColorsExample.tsx
