+++ frontend/src/common/examples/NavigationExample.tsx
