+++ frontend/src/common/examples/TableExample.tsx
