--- frontend/src/components/form/FormModal.tsx
@@ -1,2 +1,0 @@
-// this will become a reusable form modal component using the ActionModal.
-// children should be the form and it should have callback for the onsubmit and oncancel.
