+++ frontend/src/routes/(app)/_examples/demo.basic.lazy.tsx
@@ -1,5 +1,5 @@
+import { BasicExample } from "@/common/examples/BasicExample";
 import { PageContainer } from "@/components/common/PageContainer";
-import { BasicExample } from "@/examples/BasicExample";
 import { createLazyFileRoute } from "@tanstack/react-router";
 
 export const Route = createLazyFileRoute("/(app)/_examples/demo/basic")({
