+++ frontend/src/routes/(app)/_examples/demo.colors.lazy.tsx
@@ -1,5 +1,5 @@
+import { ColorsExample } from "@/common/examples/ColorsExample";
 import { PageContainer } from "@/components/common/PageContainer";
-import { ColorsExample } from "@/examples/ColorsExample";
 import { createLazyFileRoute } from "@tanstack/react-router";
 
 export const Route = createLazyFileRoute("/(app)/_examples/demo/colors")({
