+++ frontend/src/routes/(app)/_examples/demo.dialogs.tsx
@@ -1,5 +1,5 @@
+import { DialogExample } from "@/common/examples/DialogExample";
 import { PageContainer } from "@/components/common/PageContainer";
-import { DialogExample } from "@/examples/DialogExample";
 import { createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/(app)/_examples/demo/dialogs")({
