+++ frontend/src/routes/(app)/_examples/demo.forms.lazy.tsx
@@ -1,5 +1,5 @@
+import { FormsExample } from "@/common/examples/FormExample";
 import { PageContainer } from "@/components/common/PageContainer";
-import { FormsExample } from "@/examples/FormExample";
 import { createLazyFileRoute } from "@tanstack/react-router";
 
 export const Route = createLazyFileRoute("/(app)/_examples/demo/forms")({
