+++ frontend/src/routes/(app)/_examples/demo.loaders.lazy.tsx
@@ -1,5 +1,5 @@
+import { LoadersExample } from "@/common/examples/LoadersExample";
 import { PageContainer } from "@/components/common/PageContainer";
-import { LoadersExample } from "@/examples/LoadersExample";
 import { createLazyFileRoute } from "@tanstack/react-router";
 
 export const Route = createLazyFileRoute("/(app)/_examples/demo/loaders")({
