+++ frontend/src/routes/(app)/_examples/demo.tables.lazy.tsx
@@ -1,5 +1,5 @@
+import { TableExample } from "@/common/examples/TableExample";
 import { PageContainer } from "@/components/common/PageContainer";
-import { TableExample } from "@/examples/TableExample";
 import { createLazyFileRoute } from "@tanstack/react-router";
 
 export const Route = createLazyFileRoute("/(app)/_examples/demo/tables")({
