+++ frontend/src/routes/(app)/_examples/demo.typography.lazy.tsx
@@ -1,5 +1,5 @@
+import { TypographyExample } from "@/common/examples/TypographyExample";
 import { PageContainer } from "@/components/common/PageContainer";
-import { TypographyExample } from "@/examples/TypographyExample";
 import { createLazyFileRoute } from "@tanstack/react-router";
 
 export const Route = createLazyFileRoute("/(app)/_examples/demo/typography")({
